+++ packages/api/src/connections/index.ts
@@ -1,5 +1,5 @@
-import { ApiPromise, WsProvider } from '@polkadot/api';
-import { formatBalance } from '@polkadot/util';
+import { ApiPromise, HttpProvider, WsProvider } from '@polkadot/api'
+import { formatBalance } from '@polkadot/util'
 import '@subsocial/definitions/interfaces/types-lookup'
 import '@subsocial/definitions/interfaces/augment-api'
 
@@ -9,15 +9,17 @@ let api: ApiPromise | undefined = undefined
 export const getSubstrateApi = async (nodeUrl?: string) => {
   if (api) return api
 
-  const rpcEndpoint = nodeUrl || 'ws://127.0.0.1:9944/';
-  const provider = new WsProvider(rpcEndpoint);
+  const rpcEndpoint = nodeUrl || 'ws://127.0.0.1:9944/'
+  const provider = rpcEndpoint.startsWith('http')
+    ? new HttpProvider(rpcEndpoint)
+    : new WsProvider(rpcEndpoint)
 
   api = await ApiPromise.create({ provider })
 
   formatBalance.setDefaults({
     decimals: api.registry.chainDecimals,
     unit: api.registry.chainTokens
-  });
+  })
 
   return api
 }