+++ packages/api/src/types/common.ts
@@ -1,8 +1,14 @@
-import { AccountId } from '@polkadot/types/interfaces';
-import BN from 'bn.js';
-import { SpaceId, PostId, Space, Post, ReactionId } from '@subsocial/definitions/interfaces';
-import { ApiPromise as SubstrateApi } from '@polkadot/api';
-import { SubsocialIpfsProps } from '../ipfs';
+import { AccountId } from '@polkadot/types/interfaces'
+import BN from 'bn.js'
+import {
+  SpaceId,
+  PostId,
+  Space,
+  Post,
+  ReactionId
+} from '@subsocial/definitions/interfaces'
+import { ApiPromise as SubstrateApi } from '@polkadot/api'
+import { SubsocialIpfsProps } from '../ipfs'
 
 export type HttpRequestMethod = 'post' | 'get'
 
@@ -11,29 +17,34 @@ export type UseServerProps = {
 }
 
 export type SubsocialContext = {
+  /** @deprecated */
   useServer?: UseServerProps
 }
 
-export type SubsocialApiProps = SubsocialContext & SubsocialIpfsProps & {
-  substrateApi: SubstrateApi,
+export type SubsocialApiProps = SubsocialContext &
+  SubsocialIpfsProps & {
+    substrateApi: SubstrateApi
+  }
+
+export type CreateSubsocialApiProps = Omit<
+  SubsocialApiProps,
+  'substrateApi'
+> & {
+  substrateNodeUrl: string
 }
 
-export type CreateSubsocialApiProps = Omit<SubsocialApiProps, 'substrateApi'> & {
-  substrateNodeUrl: string,
-}
-
-export type SubstrateId = SpaceId | PostId | BN | string;
-export type CommonStruct = Space | Post;
-export type AnyAccountId = AccountId | string;
-export type AnySpaceId = SpaceId | BN | string;
-export type AnyPostId = PostId | BN | string;
-export type AnyReactionId = ReactionId | BN | string;
+export type SubstrateId = SpaceId | PostId | BN | string
+export type CommonStruct = Space | Post
+export type AnyAccountId = AccountId | string
+export type AnySpaceId = SpaceId | BN | string
+export type AnyPostId = PostId | BN | string
+export type AnyReactionId = ReactionId | BN | string
 
 type CidAsStr = string
 
 export type ContentResult<T> = Record<CidAsStr, T>
 
-export type PalletName = 
+export type PalletName =
   | 'domains'
   | 'permissions'
   | 'posts'
@@ -48,13 +59,13 @@ export type PalletName =
   | 'utils'
 
 export type EventsName =
-  'AccountFollowed' |
-  'SpaceFollowed' |
-  'SpaceCreated' |
-  'CommentCreated' |
-  'CommentReplyCreated' |
-  'PostCreated' |
-  'PostShared' |
-  'CommentShared' |
-  'PostReactionCreated' |
-  'CommentReactionCreated'
+  | 'AccountFollowed'
+  | 'SpaceFollowed'
+  | 'SpaceCreated'
+  | 'CommentCreated'
+  | 'CommentReplyCreated'
+  | 'PostCreated'
+  | 'PostShared'
+  | 'CommentShared'
+  | 'PostReactionCreated'
+  | 'CommentReactionCreated'