+++ CHANGELOG.md
@@ -7,22 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 0.8.2 - 2023-03-02
+## 0.8.6 - 2023-03-14
+### Added
+- improve `getContent` for IPFS instance
+- add `headers` for IPFS constructor
 
+## 0.8.5 - 2023-03-10
+### Added
+- add `properties` for `pinContent` in IPFS instance
+
+## 0.8.4 - 2023-03-10
+### Added
+- deprecated `useServer` for IPFS
+- add headers for IPFS constructor
+
+## 0.8.3 - 2023-03-09
+### Added
+- add HTTP support for connecting to Substrate node
+
+## 0.8.2 - 2023-03-02
 ### Added
 - add adminClient for IPFS instance for better management of IPFS
 
 ## 0.8.1 - 2023-02-17
-
 ### Added
 - Add flatten function for reactions
 - Add `findReaction` and `findReactions` to top-level API (wrappers around `api.blockchain.findReaction` and `api.blockchain.findReactions`)
 - Add new type `ReactionStruct`
 - Add new attribute `username` for Space in Elastic Index type
 - Add typedoc for `@subsocial/utils` package
-
-### Fixed
-- Fix behavior of `simpleFormatBalance` `currency` argument not working
 
 ### Changed
 - Change the return type of ids in `api.blockchain` functions to string
@@ -35,6 +48,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Remove `profiles` from elastic index types
+
+### Fixed
+- Fix behavior of `simpleFormatBalance` `currency` argument not working
 
 ## 0.7.13 - 2023-02-13
 ### Added